pyproject.toml
[tool.poetry]
name = "automacao_casa"
version = "0.1.0"
description = ""
authors = ["Rodrigoneal <rodrigho2006@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"pushbullet.py" = "^0.12.0"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest-sugar = "^0.9.4"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
